third_party/WebKit/LayoutTests/external/wpt/mediacapture-streams/MediaStream-default-feature-policy.https-expected.txt

This is a testharness.js-based test.
PASS Default "microphone" feature policy ["self"] allows the top-level document.
PASS Default "microphone" feature policy ["self"] allows same-origin iframes.
FAIL Default "microphone" feature policy ["self"] disallows cross-origin iframes. assert_equals: expected "#[object NavigatorUserMediaError]" but got "#OK"
PASS Feature policy "microphone" can be enabled in cross-origin iframes using "allow" attribute.
PASS Default "camera" feature policy ["self"] allows the top-level document.
PASS Default "camera" feature policy ["self"] allows same-origin iframes.
FAIL Default "camera" feature policy ["self"] disallows cross-origin iframes. assert_equals: expected "#[object NavigatorUserMediaError]" but got "#OK"
PASS Feature policy "camera" can be enabled in cross-origin iframes using "allow" attribute.
PASS Default "camera; microphone" feature policy ["self"] allows the top-level document.
PASS Default "camera; microphone" feature policy ["self"] allows same-origin iframes.
FAIL Default "camera; microphone" feature policy ["self"] disallows cross-origin iframes. assert_equals: expected "#[object NavigatorUserMediaError]" but got "#OK"
PASS Feature policy "camera; microphone" can be enabled in cross-origin iframes using "allow" attribute.
Harness: the test ran to completion.
